// src/types/writer-persistence-authorization-reconsideration-no-go.ts
export type WriterPersistenceAuthorizationReconsiderationNoGoMode =
  "persistence_adapter_implementation_authorization_reconsideration_no_go_packet_only";

export type WriterPersistenceAuthorizationReconsiderationNoGoStatus =
  | "no_go_external_evidence_missing"
  | "manual_review_blocked";

export type WriterPersistenceAuthorizationReconsiderationNoGoItem = {
  id: string;
  category: WriterPersistenceAuthorizationRemediationCategory;
  title: string;
  status: WriterPersistenceAuthorizationReconsiderationNoGoStatus;
  owner: WriterPersistenceAuthorizationRemediationOwner;
  sourcePreflightItemIds: string[];
  sourceNoGoItemIds: string[];
  sourceReviewItemIds: string[];
  sourceRemediationItemIds: string[];
  sourcePreflightStatus: WriterPersistenceAuthorizationReconsiderationPreflightStatus;
  sourceRefs: string[];
  noGoQuestion: string;
  noGoConclusion: string;
  blockingEvidence: string[];
  unresolvedPreflightGaps: string[];
  forbiddenShortcuts: string[];
  reconsiderationRequirements: string[];
  safeEscalationRefs: string[];
  redactionRules: string[];
  nonAcceptanceClauses: string[];
  nextSafeAction: string;
};

export type WriterPersistenceAuthorizationReconsiderationNoGoRuntimeFlags = {
  allRuntimeEffectsBlocked: true;
  wouldAcceptReconsiderationNoGo: false;
  wouldRecordReconsiderationNoGo: false;
  wouldDenyImplementationAuthorizationFromReconsideration: false;
  wouldPromoteToReconsiderationRemediation: false;
  wouldAcceptReconsiderationPreflight: false;
  wouldRecordReconsiderationPreflight: false;
  wouldMarkReconsiderationReady: false;
  wouldStartAuthorizationReconsideration: false;
  wouldAcceptRemediationReviewNoGo: false;
  wouldRecordRemediationReviewNoGo: false;
  wouldDenyImplementationAuthorizationFromReview: false;
  wouldPromoteToAuthorizationReconsideration: false;
  wouldAcceptRemediationReview: false;
  wouldRecordRemediationReview: false;
  wouldStoreRemediationReviewEvidence: false;
  wouldMarkExternalRemediationReviewed: false;
  wouldAcceptExternalRemediationState: false;
  wouldAcceptExternalApprovalArchive: false;
  wouldStoreApprovalArtifact: false;
  wouldUploadApprovalArtifact: false;
  wouldReadExternalArtifact: false;
  wouldHashExternalArtifact: false;
  wouldPersistArchiveIndex: false;
  wouldMarkArchiveComplete: false;
  wouldCreateAuthorizationRecord: false;
  wouldRecordAuthorizationDecision: false;
  wouldRecordAuthorizationNoGoDecision: false;
  wouldAcceptAuthorizationNoGoDecision: false;
  wouldAcceptRemediationPlan: false;
  wouldRecordRemediationEvidence: false;
  wouldMarkBlockerResolved: false;
  wouldCreateRemediationTicket: false;
  wouldDenyImplementationAuthorization: false;
  wouldGrantImplementationAuthorization: false;
  wouldRecordHumanDecision: false;
  wouldAcceptHumanDecision: false;
  wouldStoreDecisionArtifact: false;
  wouldAcceptReleaseNoGo: false;
  wouldRecordGoDecision: false;
  wouldGrantReleaseApproval: false;
  wouldEnableFeatureFlag: false;
  wouldDeployCode: false;
  wouldRunProductionWriter: false;
  wouldCollectSignature: false;
  wouldRecordOwnerApproval: false;
  wouldGrantImplementationApproval: false;
  wouldCreateApprovalRecord: false;
  wouldAcceptPatchReview: false;
  wouldReviewRealPatch: false;
  wouldAcceptPatch: false;
  wouldGeneratePatch: false;
  wouldApplyPatch: false;
  wouldModifyFiles: false;
  wouldCreateFiles: false;
  wouldDeleteFiles: false;
  wouldRunGitCommand: false;
  wouldCreateBranch: false;
  wouldCheckoutBranch: false;
  wouldCreatePullRequest: false;
  wouldCreateTestFiles: false;
  wouldRunAutomatedTests: false;
  wouldCreateImplementationPlan: false;
  wouldCreateImplementationBranch: false;
  wouldCreateAdapterCode: false;
  wouldImportRealWriterImplementation: false;
  wouldRunTransaction: false;
  wouldCreateServiceRoleClient: false;
  wouldReadServiceRoleSecret: false;
  wouldPersistEvidence: false;
  wouldStoreRawPayload: false;
  wouldStoreSecrets: false;
  wouldWriteRows: false;
  wouldWriteAuditRows: false;
  wouldReserveIdempotencyKeys: false;
  wouldWriteIdempotencyRows: false;
  wouldWriteCompensationRows: false;
  wouldCreateMigrationFile: false;
  wouldApplyMigration: false;
  wouldCreateTables: false;
  wouldEnableWriters: false;
  wouldCallAi: false;
  wouldCallStripe: false;
  wouldUnlockReports: false;
};

export type WriterPersistenceAuthorizationReconsiderationNoGoPayload =
  WriterPersistenceAuthorizationReconsiderationNoGoRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    reconsiderationNoGoMode: WriterPersistenceAuthorizationReconsiderationNoGoMode;
    sourceReconsiderationPreflightMode: WriterPersistenceAuthorizationReconsiderationPreflightMode;
    checkedAt: string;
    noGoItemCount: number;
    noGoCount: number;
    manualReviewBlockedCount: number;
    reconsiderationStillBlockedCount: number;
    blockingEvidenceCount: number;
    unresolvedPreflightGapCount: number;
    forbiddenShortcutCount: number;
    reconsiderationRequirementCount: number;
    sourcePreflightItemCount: number;
    sourceBlockedPreflightItemCount: number;
    sourceExternalEvidenceMissingCount: number;
    sourceManualReviewerRequiredCount: number;
    reconsiderationNoGoPacketReady: true;
    reconsiderationNoGoPacketOnly: true;
    sourcePreflightChecklistReady: true;
    sourcePreflightChecklistOnly: true;
    sourceReviewNoGoPacketReady: true;
    sourceReviewNoGoPacketOnly: true;
    sourceReleaseStillBlocked: true;
    preflightPassed: false;
    preflightAccepted: false;
    preflightRecorded: false;
    reconsiderationEligible: false;
    reconsiderationNoGoAccepted: false;
    reconsiderationNoGoRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationRemediationAccepted: false;
    implementationAuthorizationDecisionReady: false;
    implementationAuthorizationDecisionRecorded: false;
    implementationAuthorizationNoGoAccepted: false;
    implementationAuthorizationDenied: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalRemediationStatesAccepted: false;
    remediationReviewAccepted: false;
    remediationReviewComplete: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    implementationApprovalGranted: false;
    implementationBranchApproved: false;
    implementationPlanApproved: false;
    readyToApplyPatch: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    implementationReviewComplete: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    noGoRules: string[];
    remediationBoundaryRules: string[];
    sourceBlockedCodes: string[];
    noGoItems: WriterPersistenceAuthorizationReconsiderationNoGoItem[];
  };

export type WriterPersistenceAuthorizationReconsiderationNoGoProbeResult =
  WriterPersistenceAuthorizationReconsiderationNoGoRuntimeFlags & {
    safeMode: true;
    readOnly: true;
    blocked: true;
    reconsiderationNoGoMode: WriterPersistenceAuthorizationReconsiderationNoGoMode;
    itemId?: string;
    itemTitle?: string;
    itemStatus?: WriterPersistenceAuthorizationReconsiderationNoGoStatus;
    summary: string;
    reconsiderationNoGoPacketOnly: true;
    sourcePreflightChecklistOnly: true;
    sourceReviewNoGoPacketOnly: true;
    sourceReleaseStillBlocked: true;
    preflightAccepted: false;
    preflightRecorded: false;
    reconsiderationEligible: false;
    reconsiderationNoGoAccepted: false;
    reconsiderationNoGoRecorded: false;
    implementationAuthorizationReconsiderationReady: false;
    implementationAuthorizationGranted: false;
    implementationAuthorized: false;
    authorizationDecisionRecorded: false;
    authorizationArtifactStored: false;
    externalRemediationStatesAccepted: false;
    remediationReviewNoGoAccepted: false;
    remediationReviewNoGoRecorded: false;
    externalApprovalArchiveAccepted: false;
    archiveCompletenessAccepted: false;
    readyToCreateImplementationBranch: false;
    readyForAdapterImplementation: false;
    readyForReleaseExecution: false;
    adapterImplemented: false;
    adapterImplementationApproved: false;
    adapterImplementationAllowed: false;
    allOwnerApprovalsComplete: false;
    allBlockingEvidenceReady: false;
    blockedCodes: string[];
    noGoItems: WriterPersistenceAuthorizationReconsiderationNoGoItem[];
  };